requirements.txt
livekit==0.13.0
livekit-agents==0.13.0
livekit-plugins-silero==0.13.0
livekit-plugins-turn-detector==0.13.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pypdf==4.0.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
mcp==1.0.0
numpy==1.24.0
